engine, add support for parsing spt and dst port for iptables
At the moment we only parse ports and assume they are source port.
It is better to differentiate source and destination port for better
rules.

In this patch the Sagan_Parse_Port code is renamed in Sagan_Parse_Src_Port
because it was used to fill normalize_src_port. We add a Sagan_Parse_Dst_Port
to hold parsing of destination port.

This patch also include the parsing of spt=port and dst=port which is found
in iptables logs.