Daemonization Fixes
Change the daemonization logic according to the documentation at
https://www.svbug.com/documentation/comp.unix.programmer-FAQ/faq_2.html#SEC16.
This particularly matters when invoking 'sagan --daemon' directly, without
going through something like init.d/Upstart/systemd. As a simplified reproducible,
running "ssh some.host 'sagan --daemon ...'" would result in the SSH session
never exiting, as the SSH session connects to the child's stdout and stderr
using pipes and waits for an EOF.

Also, move the daemonization earlier in the initialization logic. By the time
sagan daemonized, it had already loaded all the rules, increasing the parent
process' memory footprint. Since a fork results in duplicating the address
space, we could get into a situation where there would be enough memory for
sagan to load all the rules and run, but not enough memory for two copies of
the sagan process, causing the fork to fail. If we fork earlier, less memory
is required, as the rules and other configuration haven't been loaded.